# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DivApi_2",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Регистрация",
                "operationId": "ad77fe138bcb384ad65d167d5c8d1cc0",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Имя пользователя",
                        "example": "User"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email пользователя",
                        "example": "test@email.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Пароль пользователя",
                        "example": "password"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Invalid data"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Авторизация",
                "operationId": "bae68d1c55329ecb6e41f48285840c8f",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email пользователя",
                        "example": "test@email.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Пароль пользователя",
                        "example": "password"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Invalid data"
                    }
                }
            }
        },
        "/api/v1/requests": {
            "get": {
                "tags": [
                    "Заявки"
                ],
                "summary": "Получение контента страницы",
                "operationId": "2169f41a4657151d729da3483d623b93",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Active = 0, Resolved = 1",
                        "example": "1"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Date in format like 2022-09-06",
                        "example": "2022-09-06"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "wrong data"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Заявки"
                ],
                "summary": "Создание заявки",
                "operationId": "26abcf9b5c3f8bdf983d32a90cad10b4",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Имя пользователя",
                        "example": "User"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email пользователя",
                        "example": "test@email.com"
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "Сообщение пользователя",
                        "example": "Test message"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Invalid data"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/requests/{id}": {
            "put": {
                "tags": [
                    "Заявки"
                ],
                "summary": "Добавление ответа к заявке",
                "operationId": "88dc3a930ddf3f8e8ac6eca525625d5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id заявки",
                        "example": 1
                    },
                    {
                        "name": "comment",
                        "in": "query",
                        "description": "Комментарий к заявке",
                        "example": "Test Comment"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Invalid data"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}